update manifest takes a glob
