+++ core/plugins/manifest/manifest.ts
@@ -1,13 +1,11 @@
 import { fs } from "$effects/fs.ts";
 import { hmr } from "$effects/hmr.ts";
 import { manifest, manifestPath } from "$effects/manifest.ts";
+import { config } from "$effects/mod.ts";
 import type { ManifestBase } from "$lib/types.d.ts";
-import { expandGlobWorkspaceRelative } from "$lib/utils/fs.ts";
-import { extractImports } from "$lib/utils/parse.ts";
 import { stringifyObject } from "$lib/utils/stringify.ts";
 import { Handler, handlerFor, type Plugin } from "@radish/effect-system";
-import type { ExpandGlobOptions } from "@std/fs";
-import { extname } from "@std/path";
+import { extname, globToRegExp } from "@std/path";
 
 let manifestObject: ManifestBase = {
   imports: {},
@@ -122,13 +120,32 @@ export const skipManifest = /\.(d|spec|test)\.(js|ts)$/;
 /**
  * Performs the manifest/update effect on all entries matching the glob
  */
-export const updateManifest = async (
-  glob: string | URL,
-  options?: ExpandGlobOptions,
-): Promise<void> => {
-  for await (const entry of expandGlobWorkspaceRelative(glob, options)) {
-    if (!skipManifest.test(entry.path)) {
-      await manifest.update(entry);
-    }
+export const updateManifest = async (glob: string): Promise<void> => {
+  const match = globToRegExp(glob);
+
+  const entries = await fs.walk(Deno.cwd(), {
+    match: [new RegExp(match.source.slice(1))],
+    includeDirs: false,
+    skip: (await config.read()).manifest?.skip ?? [/(\.test|\.spec)\.ts$/],
+  });
+
+  for await (const entry of entries) {
+    await manifest.update(entry);
   }
 };
+
+/**
+ * Extracts import specifiers from import declarations or dynamic imports
+ */
+export const import_regex =
+  /from\s["']([^'"]+)["']|import\(["']([^"']+)["']\)/g;
+
+/**
+ * Returns the deduped array of (maybe dynamic) import specifiers from a js/ts
+ * source file
+ */
+export function extractImports(source: string) {
+  return Array.from(
+    new Set(source.matchAll(import_regex).map((match) => match[1] || match[2])),
+  ).filter((str) => str !== undefined);
+}
