+++ core/utils/async.ts
@@ -15,15 +15,3 @@ export const setTimeoutWithAbort = (
     clearTimeout(id);
   });
 };
-
-/**
- * Applies an async callback to every entry of an array and returns the new array
- */
-export const asyncMap = async <T, U>(arr: T[], fn: (el: T) => Promise<U>) => {
-  const newArr = [];
-  for (const el of arr) {
-    const res = await fn(el);
-    newArr.push(res);
-  }
-  return newArr;
-};
