--- core/utils/io.ts
@@ -1,5 +1,0 @@
-export function throwUnlessNotFound(error: unknown) {
-  if (!(error instanceof Deno.errors.NotFound)) {
-    throw error;
-  }
-}
